seems to fixed binary appended raw format problem in 64 bit. it was an incorrect data type for blocklength. float instead of unsigned long corrected the problem at least
[[Imported from SVN: r3896]]

+++ grid/io/file/vtk/vtkwriter.hh
@@ -1213,7 +1213,7 @@ namespace Dune
       SimpleStream stream(s);
 
       // write length before each data block
-      unsigned long blocklength;
+      float blocklength;
 
       // point data
       for (FunctionIterator it=vertexdata.begin(); it!=vertexdata.end(); ++it)
